requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.0
joblib==1.1.0
Flask-CORS==3.0.10